fix: GraphAgent time_range_hours truncation bug
BUG: int() was truncating time_range_hours, causing incorrect time ranges
- 2.0 hours became 2 hours (works by accident)
- 1.5 hours became 1 hour (BUG!)
- 0.5 hours (30 min) became 0 hours (BUG!)

FIX: Keep as float to support minute-based searches
- Now correctly passes 2.0, 1.5, 0.5, 0.167 hours to Neo4j

+++ mini-server-prod/kgroot-services/ai-sre-platform/src/agents/graph_agent.py
@@ -54,7 +54,8 @@ class GraphAgent(BaseAgent):
                     time_window_start = datetime.fromisoformat(time_window_start.replace('Z', '+00:00'))
                 if isinstance(time_window_end, str):
                     time_window_end = datetime.fromisoformat(time_window_end.replace('Z', '+00:00'))
-                time_range_hours = int((time_window_end - time_window_start).total_seconds() / 3600)
+                # Keep as float to support minute-based searches
+                time_range_hours = (time_window_end - time_window_start).total_seconds() / 3600
             else:
                 time_range_hours = 24  # Default fallback
